+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.0.2 (Sep 11, 2017)
+
+* Add generic class support for C#.
+* Fix empty method comments for C#.
+* Remove enter command for C# comments.
+
 ## 1.0.0 (Sep 7, 2017)
 
 * Start repo.